Assignment_1/requirement.txt
streamlit==1.22.0
Pillow==9.0.1
torch==2.0.0
torchvision==0.15.1
torchmetrics